Refactor xbrl
